TRUNK-6684: Location.childLocations gained unintended orphanRemoval=true in the JPA migration
