+++ api/src/test/java/org/openmrs/LocationTest.java
@@ -10,15 +10,10 @@
 package org.openmrs;
 
 import org.junit.jupiter.api.Test;
-import org.openmrs.api.LocationService;
-import org.openmrs.api.context.Context;
-import org.openmrs.test.jupiter.BaseContextSensitiveTest;
 
-import static org.junit.jupiter.api.Assertions.assertEquals;
-import static org.junit.jupiter.api.Assertions.assertNotNull;
 import static org.junit.jupiter.api.Assertions.assertTrue;
 
-public class LocationTest extends BaseContextSensitiveTest {
+public class LocationTest {
 
 	/**
 	 * Get locations that have any of specified set of tags among its child tags.
@@ -36,33 +31,6 @@ public class LocationTest extends BaseContextSensitiveTest {
 
 		assertTrue(Location.isInHierarchy(locationChild, locationParent));
 		assertTrue(Location.isInHierarchy(locationChild, locationGrandParent));
-	}
-
-	/**
-	 * @see Location#removeChildLocation(Location)
-	 */
-	@Test
-	public void removeChildLocation_shouldNotDeleteLocationWhenReparenting() {
-		executeDataSet("org/openmrs/api/include/LocationServiceTest-initialData.xml");
-		LocationService locationService = Context.getLocationService();
-
-		Location child = locationService.getLocation(2);
-		Location oldParent = child.getParentLocation();
-		Location newParent = locationService.getLocation(3);
-
-		// Removing the child should only change the relationship, not delete the location.
-		oldParent.removeChildLocation(child);
-		child.setParentLocation(newParent);
-		locationService.saveLocation(child);
-
-		Context.flushSession();
-		Context.clearSession();
-
-		// Reload the child to verify that it still exists after the relationship change.
-		Location reloadedChild = locationService.getLocation(child.getLocationId());
-
-		assertNotNull(reloadedChild);
-		assertEquals(newParent.getLocationId(), reloadedChild.getParentLocation().getLocationId());
 	}
 
 }
